refactor to move data validation and util functions from controllers to models
