+++ controllers/articles.controllers.js
@@ -3,7 +3,6 @@ const {
 	fetchArticles,
 	updateArticleByArticleId,
 } = require("../models/articles.models");
-const { checkTopicExists } = require("./controllers.utils");
 
 exports.getArticleById = (req, res, next) => {
 	const { article_id } = req.params;
@@ -18,18 +17,9 @@ exports.getArticleById = (req, res, next) => {
 
 exports.getArticles = (req, res, next) => {
 	const { topic, sort_by, order } = req.query;
-	const articlesQuery = fetchArticles(topic, sort_by, order);
-	const queries = [articlesQuery];
-	if (topic) {
-		const topicExistsQuery = checkTopicExists(topic);
-		queries.push(topicExistsQuery);
-	}
-	Promise.all(queries)
-		.then((resolvedPromises) => {
-			const articles = resolvedPromises[0];
-			articles.forEach((article) => {
-				return delete article.body;
-			});
+
+	fetchArticles(topic, sort_by, order)
+		.then((articles) => {
 			res.status(200).send({ articles: articles });
 		})
 		.catch((err) => {
