+++ controllers/comments.controllers.js
@@ -3,16 +3,11 @@ const {
 	insertCommentByArticleId,
 	removeCommentById,
 } = require("../models/comments.models");
-const { checkArticleExists } = require("./controllers.utils");
 
 exports.getCommentsByArticleId = (req, res, next) => {
 	const { article_id } = req.params;
-	const fetchCommentsQuery = fetchCommentsByArticleId(article_id);
-	const articleExistsQuery = checkArticleExists(article_id);
-
-	Promise.all([fetchCommentsQuery, articleExistsQuery])
-		.then((response) => {
-			const comments = response[0];
+	fetchCommentsByArticleId(article_id)
+		.then((comments) => {
 			res.status(200).send({ comments });
 		})
 		.catch((err) => {
